arm64: Load FDT at the end of DRAM with firmware
We load the FDT right after the firmware. This is problematic, if the firmware
image has sufficiently larger reserved uninitialised data at the end of the image
(e.g., bss, stack), as the FDT may be overwritten at init, before the image gets
a chance to relocate it.

Fix this by following the same policy as the kernel, where the FDT is loaded at
the end of the DRAM.

Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>